Add files via upload
Agriculture is the main source for growth and development of economy and a plays a crucial role in variety of industrial sectors. Unfortunately, farmers are still dependent on traditional methods which has poor yields and other drawbacks. Modern day farming demands increased production of food to accommodate the large global population. The alarming climate change and scarcity of water demand improved methods for agriculture. There is a need to shift towards automated farming systems based on sensors which can be used to integrate informations corresponding to the ideal farming conditions and to use the output data for analysis and management. This sensor based system contains channels for responding to the sensor outputs and can improve crop growth from the farms. This proposed system consists of sensors like soil moisture sensor, humidity sensor, pH sensor and temperature sensor. These sensors collects the present soil moisture, humidity, soil pH and temperature of the surrounding. The project aims at integrating traditional farming and modern techniques to reduce the manual monitoring of soil testing and irrigation and help farmers in increasing the agricultural production. The main purpose of this system is to increase the agricultural yield and reduces the investments in soil monitoring and irrigation.